[test_nn] split lazy_modules from test_nn
